membuat struktur folder halaman kriteria dan membuat fitur tambah data

+++ database/migrations/2025_07_07_025831_create_kriteria_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('tb_kriteria', function (Blueprint $table) {
             $table->id();
             $table->string('nama_kriteria', 50);
-            $table->float('bobot', 20);
+            $table->double('bobot');
             $table->enum('atribut', ['benefit', 'cost']);
             $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
